Updating the advanced documation.
This commit add a note on the use of km vs. planetary radius for the spatial parameters. If OBSERVER TYPE Pre-Defined with EPHEM: file.csv, all spatial parameters need to be in km (*not* in planetary radius).
The conversion is then done by the code itself in read_save.pro